# Conflicts:
#	translations/client_de.ts
#	translations/client_es.ts
#	translations/client_fr.ts
#	translations/client_it.ts

+++ src/gui/customtabbar.cpp
@@ -1,6 +1,6 @@
 /*
  * Infomaniak kDrive - Desktop
- * Copyright (C) 2023-2025 Infomaniak Network SA
+ * Copyright (C) 2023-2026 Infomaniak Network SA
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -34,7 +34,7 @@ static int textLineHeightOffset = 15;
 static int notifBoxXMargin = 10;
 static int notifBoxHeight = 20;
 static float notifBoxRadius = 9.5;
-static int maxNbNotif = 99;
+static uint16_t maxNbNotif = 99;
 static int notifTextPadding = 15;
 
 CustomTabBar::CustomTabBar(QWidget *parent) :
